requirements.txt
langchain
boto3
faiss-cpu
python-dotenv
requests
